Created Account, Transfer and other improved

+++ src/main/java/interfaces/BankAccount.java
@@ -3,9 +3,9 @@ package main.java.interfaces;
 public interface BankAccount {
     String getAccNumber();
 
-    double getBalance();
+    int getBalance();
 
-    void setBalance(double balance);
+    void setBalance(int balance);
 
     void setAccNumber(String accNumber);
 }
